+++ src/libs/components/pagination/pagination.module.ts
@@ -9,7 +9,16 @@ import { NtPaginationComponent } from './pagination.component';
   declarations: [NtPaginationComponent]
 })
 export class NtPaginationModule {
-  public static forRoot(config?: NtPaginationConfig): ModuleWithProviders {
+
+  /**
+   * 此方法将会废弃，请用 NtPaginationModule.withConfig 代替
+   * @deprecated 0.4.x 废弃
+   */
+  public static forRoot(config?: NtPaginationConfig) {
+    return NtPaginationModule.withConfig(config);
+  }
+
+  public static withConfig(config?: NtPaginationConfig): ModuleWithProviders {
     return {
       ngModule: NtPaginationModule,
       providers: [